fix(stock): completar la migración a stock decimal en el código
La migración 32 pasó las columnas de stock a numeric(12,2), pero quedaron
dos frentes sin ajustar y los dos venían fallando en silencio.

parseInt -> parseFloat en todo lo que carga cantidades. El diálogo de
entrada/salida (el que usan los botones de toda la app) truncaba "2.5" a 2
y encima tenía step="1", así que el navegador ni siquiera dejaba tipear la
coma. Mismo arreglo en el formulario de ajuste, en el importador de
planillas y en el stock inicial de un material nuevo. min_stock/max_stock
siguen enteros a propósito: son umbrales de alerta, no cantidades.

::float8 en los cuatro endpoints de ai-tools. El driver devuelve numeric
como string, y como las filas se serializan tal cual, el bot recibía
"1.00" donde esperaba 1; cualquier cuenta del otro lado daba mal.

De paso, los tipos de current_stock/available_stock pasan a
number | string, que es lo que realmente llega. Eso destapó una
comparación que operaba sobre el string en el diálogo.

Verificado en local: retirar 2,5 deja el movimiento en 2.50 y el stock en
314,50; antes el formulario no lo aceptaba.

+++ components/inventory-importer.tsx
@@ -49,7 +49,10 @@ export function InventoryImporter() {
             const row = line.split(/,(?=(?:(?:[^"]*"){2})*[^"]*$)/).map(c => c.trim().replace(/"/g, ''))
 
             const barcode = row[barcodeIndex]
-            const stock = parseInt(row[stockIndex])
+            // parseFloat: el stock es numeric(12,2). Con parseInt, importar una
+            // planilla con "12,5 m de cable" cargaba 12 y la diferencia no
+            // aparecía por ningún lado.
+            const stock = parseFloat(row[stockIndex])
             const cost = costIndex !== -1 ? parseFloat(row[costIndex]) : undefined
 
             if (barcode && !isNaN(stock)) {
